@@ -5,7 +5,6 @@ This module contains models for rankings.
 from __future__ import annotations
 
 from typing import Literal
-from typing import Optional
 
 from .beatmap import Beatmapset
 from .common import CursorModel
@@ -26,6 +25,6 @@ RankingVariant = Literal["4k", "7k"]
 
 class Rankings(CursorModel):
     ranking: list[UserStats]
-    total: Optional[int] = None
-    spotlight: Optional[Spotlight] = None
-    beatmapsets: Optional[list[Beatmapset]] = None
+    total: int | None = None
+    spotlight: Spotlight | None = None
+    beatmapsets: list[Beatmapset] | None = None
