@@ -5,8 +5,6 @@ This module contains models for the search endpoint.
 from __future__ import annotations
 
 from typing import Literal
-from typing import Optional
-from typing import Union
 
 from .base import BaseModel
 from .user import User
@@ -22,10 +20,10 @@ SearchMode = Literal["all", "user", "wiki_page"]
 
 
 class SearchResult(BaseModel):
-    data: list[Union[User, WikiPage]]
+    data: list[User | WikiPage]
     total: int
 
 
 class SearchResponse(BaseModel):
-    users: Optional[SearchResult] = None
-    wiki_pages: Optional[SearchResult] = None
+    users: SearchResult | None = None
+    wiki_pages: SearchResult | None = None
