@@ -4,8 +4,6 @@ This module contains models for wiki objects.
 
 from __future__ import annotations
 
-from typing import Optional
-
 from .base import BaseModel
 
 __all__ = ("WikiPage",)
@@ -18,5 +16,5 @@ class WikiPage(BaseModel):
     available_locales: list[str]
     layout: str
     markdown: str
-    subtitle: Optional[str] = None
-    tags: Optional[list[str]] = None
+    subtitle: str | None = None
+    tags: list[str] | None = None
